[i74] fix slider jump after seekTo process

+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/feature/messages/list/adapter/view/internal/AudioRecordsGroupView.kt
@@ -132,7 +132,8 @@ public class AudioRecordsGroupView : LinearLayoutCompat {
         }
         onProgressStateChange(hashCode) { (duration, progress) ->
             playerView.setDuration(DurationParser.durationInMilliToReadableTime(duration))
-            playerView.setProgress(progress)
+            // TODO
+            playerView.setProgress(progress.toDouble())
         }
         onSpeedChange(hashCode, playerView::setSpeedText)
     }
